add unit test for client propose tx
